+++ storage/fileservice/scanner.go
